ego-consciousness-uploader/elevenlabs/requirements.txt
elevenlabs>=0.2.26
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pygame>=2.0.0